% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metasat.R
\name{RMSAT}
\alias{RMSAT}
\title{Robust heterogeneous-effects meta-analysis of variant-set association}
\usage{
RMSAT(Us, Rs, eta = NULL, rho = c((0:5/10)^2, 0.5, 1))
}
\arguments{
\item{Us}{matrix of variant test statistics (m by K)}

\item{Rs}{array of covariance matrix for test statistics (m,m by K)}

\item{eta}{coefficient vector for the SBT. Default to equal weights.}

\item{rho}{weights assigned to the SBT}
}
\value{
\describe{
  \item{p.value}{ p-value for AT }
  \item{pval}{ vector of p-values for each rho }
  \item{rho.est}{ optimal rho value leading to the minimum p-value }
}
}
\description{
Conduct meta-analysis of variant-set association test of m variants assuming heterogeous effects across K studies.
SST: \eqn{Q=\sum_k U_k^TU_k}; SBT: \eqn{\sum_k(\eta_k^TU_k)^2}
}
\examples{
K = 3; m=10
Rs = array(0, dim=c(m,m,K)); Us = matrix(0, m,K)
for(k in 1:K){
  ak = matrix(rnorm(100*m),100,m)*sqrt(0.8)+rnorm(100)*sqrt(0.2)
  Rs[,,k] = cor(ak)
  Rh = chol(Rs[,,k])
  Us[,k] = colSums(Rh*rnorm(m))
}
RMSAT(Us,Rs)
U1 = Us + rnorm(m*K,1,1.25)
RMSAT(U1,Rs)
}
\references{
Guo,B., Massoti,M. and Wu,B. (2018). Efficient and powerful meta-analysis of variant-set association using MetaSAT.
}
\keyword{SAT}
