% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metasat.R
\name{HMSAT}
\alias{HMSAT}
\title{Heterogeneous-effects meta-analysis of variant-set association}
\usage{
HMSAT(Us, Rs, eta = NULL, rho = c((0:5/10)^2, 0.5, 1))
}
\arguments{
\item{Us}{matrix of variant test statistics (m, K)}

\item{Rs}{array of covariance matrix for test statistics (m, m, K)}

\item{eta}{coefficient vector for the GBT. Default to equal weights.}

\item{rho}{weights assigned to the GBT}
}
\value{
\describe{
  \item{p.value}{ p-value for the adaptive test (AT) }
  \item{pval}{ vector of p-values for each rho }
  \item{rho.est}{ optimal rho value leading to the minimum p-value }
}
}
\description{
Conduct meta-analysis of variant-set association test of m variants assuming heterogeous effects across K studies.
SST: \eqn{Q=\sum_k U_k^TU_k}; GBT: \eqn{(\sum_k\eta_k^TU_k)^2}; AT: adaptively weighting SST and GBT.
}
\examples{
K = 3; m=10
Rs = array(0, dim=c(m,m,K)); Us = matrix(0, m,K)
for(k in 1:K){
  ak = matrix(rnorm(100*m),100,m)*sqrt(0.8)+rnorm(100)*sqrt(0.2)
  Rs[,,k] = cor(ak)
  Rh = chol(Rs[,,k])
  Us[,k] = colSums(Rh*rnorm(m))
}
HMSAT(Us,Rs)
U1 = Us + rnorm(m*K)
HMSAT(U1,Rs)
}
\references{
Guo,B., Massoti,M., Liu,N. and Wu,B. (2018). Efficient and powerful meta-analysis of variant-set association using MetaSAT.
}
\keyword{SAT}
